transform to go-module project
