+++ go.mod
@@ -1,0 +1,9 @@
+module github.com/rainu/rocketchat-user-proxy
+
+require (
+	github.com/alexflint/go-arg v1.0.0
+	github.com/gorilla/handlers v1.4.0
+	github.com/gorilla/mux v1.7.0
+	github.com/gorilla/websocket v1.4.0
+	github.com/pkg/errors v0.8.1
+)
