+++ internal/messages/get_rooms.go
